ch23 2

+++ exercises/23_conversions/using_as.rs
@@ -4,8 +4,8 @@
 
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
-    // TODO: Make a conversion before dividing.
-    total / values.len()
+    // 将 values.len() 转换为 f64 类型
+    total / values.len() as f64
 }
 
 fn main() {
